split out operation to it's own file
